+++ cadl-tests/src/main/java/com/cadl/odata/OdataClient.java
@@ -43,9 +43,9 @@ public final class OdataClient {
      *     <tr><th>Name</th><th>Type</th><th>Required</th><th>Description</th></tr>
      *     <tr><td>filter</td><td>String</td><td>No</td><td>The filter parameter</td></tr>
      *     <tr><td>orderby</td><td>String</td><td>No</td><td>The orderBy parameter</td></tr>
-     *     <tr><td>skip</td><td>Long</td><td>No</td><td>The skip parameter</td></tr>
-     *     <tr><td>top</td><td>Long</td><td>No</td><td>The top parameter</td></tr>
-     *     <tr><td>maxpagesize</td><td>Long</td><td>No</td><td>The maxPageSize parameter</td></tr>
+     *     <tr><td>skip</td><td>Integer</td><td>No</td><td>The skip parameter</td></tr>
+     *     <tr><td>top</td><td>Integer</td><td>No</td><td>The top parameter</td></tr>
+     *     <tr><td>maxpagesize</td><td>Integer</td><td>No</td><td>The maxPageSize parameter</td></tr>
      *     <tr><td>select</td><td>List&lt;String&gt;</td><td>No</td><td>The select parameter. In the form of "," separated string.</td></tr>
      *     <tr><td>expand</td><td>List&lt;String&gt;</td><td>No</td><td>The expand parameter. In the form of "," separated string.</td></tr>
      * </table>
@@ -103,9 +103,9 @@ public final class OdataClient {
     public PagedIterable<Resource> list(
             String filter,
             String orderBy,
-            Long skip,
-            Long top,
-            Long maxPageSize,
+            Integer skip,
+            Integer top,
+            Integer maxPageSize,
             List<String> select,
             List<String> expand) {
         // Generated convenience method for list
